Added some validation in userService functions

+++ services/userService.js
@@ -13,29 +13,40 @@ class UserService {
     }
 
     createUser(input) {
-        const newUser = UserRepository.create(input);
-        if(!newUser) {
-            return null;
+        const { phoneNumber, email } = input;
+
+        if (!!this.search({ phoneNumber }) || !!this.search({ email })) {
+            throw new Error("Cannot add new user. There is user with such email / phone number.");
         }
 
+        const newUser = UserRepository.create(input);
         return newUser;
     }
 
-    updateUser(id, dataToChange) {
-        const updated = UserRepository.update(id, dataToChange);
+    updateUser(id, ChangedData) {
+        if (!this.search({ id })) {
+            throw new Error("Cannot update. User not found.");
+        }
+
+        if (this.search({ email: ChangedData?.email }) || this.search({ phoneNumber: ChangedData?.phoneNumber })) {
+            throw new Error("Cannot update data. There is user with such email / phone number.");
+        }
+
+        const updated = UserRepository.update(id, ChangedData);
+
         if(!updated) {
-            return null;
+            throw new Error("Cannot update. User not found.");
         }
         
         return updated;
     }
 
     deleteUser(id) {
-        const deleted = UserRepository.delete(id);
-        if(deleted.length === 0) {
-            return null;
+        if (!this.search({ id })) {
+            throw new Error("Cannot delete. User not found.");
         }
 
+        const deleted = UserRepository.delete(id);
         return deleted;
     }
